Update Springboot final test score, update new logic to read double score

+++ back/hackathon/src/main/java/com/hackathon/entity/MockData.java
@@ -27,7 +27,7 @@ public class MockData {
     private String email;
 
     @Column(nullable = false, unique = false)
-    private Integer score;
+    private Double score;
 
     public int getNo() {
         return no;
@@ -37,7 +37,7 @@ public class MockData {
         return email;
     }
 
-    public Integer getScore() {
+    public Double getScore() {
         return score;
     }
 
@@ -50,7 +50,7 @@ public class MockData {
         this.email = email;
     }
 
-    public void setScore(Integer score) {
+    public void setScore(Double score) {
         this.score = score;
     }
